Drawer successfully modularized
