+++ App.js
@@ -10,20 +10,21 @@ import {
   Button,
   ScrollView,
   LogBox,
-
-} from 'react-native';
-import MapboxGL from '@react-native-mapbox-gl/maps';
-import firestore from '@react-native-firebase/firestore';
-import Geolocation from 'react-native-geolocation-service';
-import { MAPBOXGL_ACCESS_TOKEN } from './secrets';
-import BottomSheet from 'reanimated-bottom-sheet';
-import Constants from 'expo-constants';
-import { browse } from './foursquare';
-import renderAnnotation from './renderAnnotation';
+  FlatList,
+} from "react-native";
+import MapboxGL from "@react-native-mapbox-gl/maps";
+import firestore from "@react-native-firebase/firestore";
+import Geolocation from "react-native-geolocation-service";
+import { MAPBOXGL_ACCESS_TOKEN } from "./secrets";
+import BottomSheet from "reanimated-bottom-sheet";
+import { browse } from "./foursquare";
+import renderAnnotation from "./renderAnnotation";
+import {renderInner, renderHeader} from "./drawer";
 
 LogBox.ignoreAllLogs(); //Ignore all log notifications
 
 MapboxGL.setAccessToken(MAPBOXGL_ACCESS_TOKEN);
+
 
 class App extends React.Component {
   constructor(props) {
@@ -33,6 +34,7 @@ class App extends React.Component {
       locations: [],
       permissionsGranted: null,
       foursquare: [],
+      loading: false
     };
 
     this.getFbVenues = this.getFbVenues.bind(this);
@@ -143,35 +145,12 @@ class App extends React.Component {
       };
     });
   }
-  renderInner = () => (
-    <View style={styles.panel}>
-      <Text style={styles.panelTitle}>Swipe Up To Explore!</Text>
-      {/* <Text style={styles.panelSubtitle}>So Much</Text> */}
-      <Image style={styles.photo} source={require("./assets/wakeupcat.jpg")} />
-      <ScrollView style={styles.scrollView}>
-        <Text style={styles.scrollText}>First Text Box</Text>
-        <Text style={styles.scrollText}>Second Text Box</Text>
-        <Text style={styles.scrollText}>Third Text Box</Text>
-        <Text style={styles.scrollText}>Fourth Text Box</Text>
-        <Text style={styles.scrollText}>Fifth Text Box</Text>
-        <Text style={styles.scrollText}>Sixth Text Box</Text>
-        <Text style={styles.scrollText}>Seventh Text Box</Text>
-        <Text style={styles.scrollText}>Eighth Text Box</Text>
-        <Text style={styles.scrollText}>Ninth Text Box</Text>
-        <Text style={styles.scrollText}>Tenth Text Box</Text>
-      </ScrollView>
-    </View>
-  );
 
-  renderHeader = () => (
-    <View style={styles.header}>
-      <View style={styles.panelHeader}>
-        <View style={styles.panelHandle} />
-      </View>
-    </View>
-  );
+
+
 
   myRef = React.createRef();
+
   render() {
     const venuesArray = this.state.foursquare;
     return (
@@ -208,73 +187,13 @@ class App extends React.Component {
         <BottomSheet
           ref={this.myRef}
           snapPoints={[800, 125]}
-          renderContent={this.renderInner}
-          renderHeader={this.renderHeader}
+          renderHeader={renderHeader}
+          renderContent={() => renderInner(this.state.foursquare)}
           initialSnap={1}
         />
       </View>
     );
   }
 }
-
-const styles = StyleSheet.create({
-  panelContainer: {
-    position: "absolute",
-    top: 0,
-    bottom: 0,
-    left: 0,
-    right: 0,
-  },
-  panel: {
-    height: 800,
-    padding: 20,
-    backgroundColor: "#f7f5eee8",
-  },
-  header: {
-    backgroundColor: "#f7f5eee8",
-    shadowColor: "#000000",
-    paddingTop: 20,
-    borderTopLeftRadius: 20,
-    borderTopRightRadius: 20,
-  },
-  panelHeader: {
-    alignItems: "center",
-  },
-  panelHandle: {
-    width: 40,
-    height: 8,
-    borderRadius: 4,
-    backgroundColor: "#00000040",
-    marginBottom: -10,
-  },
-  panelTitle: {
-    fontSize: 20,
-    height: 35,
-    textAlign: "center",
-  },
-  panelSubtitle: {
-    fontSize: 14,
-    color: "gray",
-    height: 30,
-    marginTop: 30,
-    marginBottom: 10,
-  },
-  photo: {
-    width: "100%",
-    height: 300,
-    marginTop: 50,
-  },
-  scrollContainer: {
-    flex: 1,
-    marginTop: Constants.statusBarHeight,
-  },
-  scrollView: {
-    backgroundColor: "pink",
-    marginHorizontal: 20,
-  },
-  scrollText: {
-    fontSize: 42,
-  },
-});
 
 export default App;
