+++ package.json
@@ -28,12 +28,14 @@
     "react-native": "~0.63.3",
     "react-native-axios": "^0.17.1",
     "react-native-camera": "git+https://git@github.com/react-native-community/react-native-camera.git",
+    "react-native-elements": "^3.0.0-alpha.1",
     "react-native-geolocation-service": "^5.1.1",
     "react-native-gesture-handler": "^1.7.0",
     "react-native-reanimated": "^1.13.2",
     "react-native-safe-area-context": "^3.1.9",
     "react-native-screens": "^2.10.1",
     "react-native-unimodules": "~0.11.0",
+    "react-native-vector-icons": "^7.1.0",
     "react-native-web": "~0.13.12",
     "reanimated-bottom-sheet": "^1.0.0-alpha.22"
   },
